requirements.txt
# Refinitiv Data API
refinitiv-data>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Database
psycopg2-binary>=2.9.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
